+++ src/utils/isSubdirectory.ts
@@ -1,5 +1,13 @@
-import { relative } from 'path';
+import path from 'path';
 
+/**
+ * Checks whether `child` is a subdirectory of `parent`
+ * @param parent Parent directory (absolute path)
+ * @param child Child directory (absolute or relative path)
+ */
 export function isSubDirectory(parent: string, child: string) {
-    return relative(child, parent).startsWith('..');
+    const parentNormalized = path.resolve(parent);
+    const childNormalized = path.resolve(parentNormalized, child);
+    const relative = path.relative(parentNormalized, childNormalized);
+    return relative !== '' && !relative.startsWith('..') && !path.isAbsolute(relative);
 }